Relocated enums and updated imports accordingly
