+++ enums/categories.js
